requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==2.3.3
Flask-WTF==1.1.1
fonttools==4.53.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pygame==2.6.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==2.3.7
WTForms==3.1.2
